Regenerator complains that it cannot find `fs` module.
Ajv introduced async support since version 3.5.0. Ajv compiles
asynchronous schemas to either generator function (default) that can be
optionally transpiled with regenerator or to es7 async function that can
be transpiled with nodent or with regenerator as well.

Regenerator was throwing a fatal error because it didn't find `fs`
module when it was compiled using `webpack`.