Clarify validation error message
When running validate in sandbox, it was not intuitive to me that it also validates the schema. I expected it to run my validate function and return only those results. My getSettings function was returning a number when a string was expected so the error was correct, but this took way too long to trouble-shoot. I think a more detailed error message can better direct users to the source of the problem.